Fix violations of a C++11 type narrowing rule
Casting from int to unsigned int or unsigned long is considered narrowing for non-const expressions in initializer lists, which is a compilation error with xlc++.

+++ include/lbann/data_readers/lbann_data_reader_nci.hpp
@@ -66,7 +66,7 @@ class data_reader_nci : public generic_data_reader {
     return m_num_labels;
   }
   const std::vector<int> get_data_dims(void) const {
-    return {m_num_features};
+    return {static_cast<int>(m_num_features)};
   }
 
  private:
